add a switch to indie button in respawn menu
Team switching is discouraged as it often can be abused for team killing
and/or easy gearing up. However, many players at times want to play with
their friends who can't always join the bluefor/opfor team they may be
on. An option to switch to Indie ONLY from the respawn menu is a good
way to trust the switch is more likely for proper intention. Once
switched player is locked to Indie, with no option to get back to
bluefor/opfor.